% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_run_speval.R
\name{shiny_run_speval}
\alias{shiny_run_speval}
\title{Run the species name validation tool}
\usage{
shiny_run_speval(...)
}
\arguments{
\item{...}{arguments to pass to \code{shinyApp()}}
}
\description{
Starts a R Shiny application that let users upload a list of species names,
check them for typos and validate them against several major taxonomic backbones.
}
\examples{
if (interactive()) {

shiny_run_speval()

}

}
